+++ src/api/todos-with-users.ts
@@ -1,0 +1,16 @@
+import todosFromServer from './todos';
+import usersFromServer from './users';
+
+import { User } from '../types/User';
+import { Todo } from '../types/Todo';
+
+function getUser(userId: number): User | null {
+  const foundUser = usersFromServer.find(user => user.id === userId);
+
+  return foundUser || null;
+}
+
+export const todos: Todo[] = todosFromServer.map(todo => ({
+  ...todo,
+  user: getUser(todo.userId),
+}));
